moved game to certain package
