+++ src/main/java/database/game.csv
@@ -2,3 +2,4 @@ user1_username,user1_truth1,user1_truth2,user1_lie,user2_username,user2_truth1,u
 sunny,Truth,Truth,Lie,leo,null,null,null
 sunny,null,null,null,queen,Truth,Truth,Lie
 sunny,null,null,null,tiffany,null,null,null
+sunny,null,null,null,ian,null,null,null
