+++ src/main/java/gui/game/TwoTruthsAndALiePageScrollPane.java
@@ -1,4 +1,7 @@
-package gui;
+package gui.game;
+import gui.MainFrame;
+import gui.game.TwoTruthsAndALiePagePanel;
+
 import javax.swing.*;
 import java.awt.*;
