+++ src/main/java/presenter/TwoTruthsAndALiePagePresenter.java
@@ -3,7 +3,7 @@ package presenter;
 import database.csvManager;
 import entity.TwoTruthsAndALieGame;
 import entity.User;
-import gui.TwoTruthsAndALiePagePanelInterface;
+import gui.game.TwoTruthsAndALiePagePanelInterface;
 import use_case_signin_signup.UserRequestModel;
 import use_case_two_truths_and_a_lie.*;
